+++ sources/packages/web/src/assets/css/global-style-variables.scss
@@ -1,6 +1,6 @@
 /** Brand Colors - Figma **/
 $brand-blue: #2965c5;
-$brand-grey: #8692a4;
+$brand-grey: #69768C;
 $brand-black: #333a47;
 $brand-black-light: #999ca3;
