+++ sources/packages/web/src/components/common/notes/Notes.vue
@@ -1,70 +1,75 @@
 <template>
-  <v-row class="m-2">
-    <v-col class="category-header-large color-blue">{{ title }}</v-col>
-    <v-col
-      ><v-btn
-        @click="addNewNote()"
-        class="float-right"
-        color="primary"
-        prepend-icon="fa:far fa-edit"
+  <full-page-container
+    :layout-template="LayoutTemplates.CenteredCard"
+    :fullWidth="true"
+  >
+    <v-row class="m-2">
+      <v-col class="category-header-large color-blue">{{ title }}</v-col>
+      <v-col
+        ><v-btn
+          @click="addNewNote()"
+          class="float-right"
+          color="primary"
+          prepend-icon="fa:far fa-edit"
+        >
+          Create new note</v-btn
+        ></v-col
       >
-        Create new note</v-btn
-      ></v-col
+    </v-row>
+    <v-row class="m-2" v-if="!notes || notes.length === 0"
+      ><v-col
+        >No notes found. Please click on create new note to add one.</v-col
+      ></v-row
     >
-  </v-row>
-  <v-row class="m-2" v-if="!notes || notes.length === 0"
-    ><v-col
-      >No notes found. Please click on create new note to add one.</v-col
-    ></v-row
-  >
-  <v-timeline
-    truncate-line="both"
-    side="end"
-    align="left"
-    class="justify-content-start"
-  >
-    <v-timeline-item
-      v-for="note in notes"
-      :key="note"
-      dot-color="primary"
-      size="x-small"
-      fill-dot
+    <v-timeline
+      truncate-line="both"
+      side="end"
+      align="left"
+      class="justify-content-start"
     >
-      <div class="d-flex">
-        <span class="primary-color marker-text">{{
-          dateOnlyLongString(note.createdAt)
-        }}</span>
-        <div class="mx-8">
-          <div class="content-header">{{ note.noteType }}</div>
-          <div v-if="showMoreNotes(notes)" class="header mt-2">
-            {{ note.description.substring(0, 150) }}
-            <a @click="toggleNotes(notes)" class="primary-color"
-              >Show more...</a
-            >
-          </div>
-          <div v-else class="header mt-2">
-            {{ note.description }}
-            <a
-              v-if="note.showMore"
-              @click="toggleNotes(notes)"
-              class="primary-color"
-              >Show less...</a
-            >
-          </div>
-          <div class="content-footer mt-2 mb-8 secondary-color-light">
-            <span>{{ timeOnlyInHoursAndMinutes(note.createdAt) }}</span
-            ><span class="mx-2">|</span>
-            <span>{{ `${note.lastName}, ${note.firstName}` }}</span>
+      <v-timeline-item
+        v-for="note in notes"
+        :key="note"
+        dot-color="primary"
+        size="x-small"
+        fill-dot
+      >
+        <div class="d-flex">
+          <span class="primary-color marker-text">{{
+            dateOnlyLongString(note.createdAt)
+          }}</span>
+          <div class="mx-8">
+            <div class="content-header">{{ note.noteType }}</div>
+            <div v-if="showMoreNotes(notes)" class="header mt-2">
+              {{ note.description.substring(0, 150) }}
+              <a @click="toggleNotes(notes)" class="primary-color"
+                >Show more...</a
+              >
+            </div>
+            <div v-else class="header mt-2">
+              {{ note.description }}
+              <a
+                v-if="note.showMore"
+                @click="toggleNotes(notes)"
+                class="primary-color"
+                >Show less...</a
+              >
+            </div>
+            <div class="content-footer mt-2 mb-8 secondary-color-light">
+              <span>{{ timeOnlyInHoursAndMinutes(note.createdAt) }}</span
+              ><span class="mx-2">|</span>
+              <span>{{ `${note.lastName}, ${note.firstName}` }}</span>
+            </div>
           </div>
         </div>
-      </div>
-    </v-timeline-item>
-  </v-timeline>
-  <CreateNoteModal
-    ref="createNotesModal"
-    :entityType="entityType"
-    @submitData="emitToParent"
-  />
+      </v-timeline-item>
+    </v-timeline>
+    <CreateNoteModal
+      ref="createNotesModal"
+      :entityType="entityType"
+      @submitData="emitToParent"
+    />
+  </full-page-container>
 </template>
 
 <script lang="ts">
@@ -73,6 +78,7 @@ import CreateNoteModal from "@/components/common/notes/CreateNoteModal.vue";
 import { NoteBaseDTO, NoteDTO } from "@/types";
 import { ref } from "vue";
 import "@/assets/css/notes.scss";
+import { LayoutTemplates } from "@/types";
 
 export default {
   components: { CreateNoteModal },
@@ -121,6 +127,7 @@ export default {
       emitToParent,
       toggleNotes,
       showMoreNotes,
+      LayoutTemplates,
     };
   },
 };
