+++ sources/packages/web/src/components/generic/StatusChipActiveUser.vue
@@ -3,7 +3,7 @@
 </template>
 <script lang="ts">
 import { computed } from "vue";
-import StatusChip from "@/components/generic/StatusChip.vue";
+import StatusChip from "@/components/generic/ChipStatus.vue";
 import { StatusChipTypes } from "@/components/generic/StatusChip.models";
 export default {
   components: { StatusChip },
