+++ sources/packages/web/src/components/institutions/active-application/ActiveApplicationSummaryData.vue
@@ -1,7 +1,7 @@
 <template>
   <full-page-container
     :layout-template="LayoutTemplates.CenteredCard"
-    :full-width="true"
+    :fullWidth="true"
   >
     <body-header
       title="Applications"
