+++ sources/packages/web/src/views/aest/institution/Designation.vue
@@ -1,7 +1,7 @@
 <template>
   <full-page-container
     :layout-template="LayoutTemplates.CenteredCard"
-    :full-width="true"
+    :fullWidth="true"
   >
     <body-header
       title="Designation agreements"
