+++ sources/packages/web/src/views/aest/institution/Restrictions.vue
@@ -1,7 +1,7 @@
 <template>
   <full-page-container
     :layout-template="LayoutTemplates.CenteredCard"
-    :full-width="true"
+    :fullWidth="true"
   >
     <body-header title="All Restrictions" class="m-1">
       <template #actions
