+++ sources/packages/web/src/views/institution/locations/confirmation-of-enrollment/COESummaryData.vue
@@ -1,7 +1,7 @@
 <template>
   <full-page-container
     :layout-template="LayoutTemplates.CenteredCard"
-    :full-width="true"
+    :fullWidth="true"
   >
     <body-header
       :title="header"
